requirements.txt
asgiref==3.3.1
Cerberus==1.3.2
Django==3.1.5
djangorestframework==3.12.2
mysqlclient==2.0.3
pytz==2021.1
SQLAlchemy==1.3.22
sqlparse==0.4.1
